finance-app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dateutil==2.9.0
yfinance
pandas